Add a Series class and scan navigation
The Series class is used to identify and validate the image files in
the series directory. It is also used to obtain the file name for
a specific row, column, and scan number.

The user may also use the arrow keys on their keyboard to navigate
through different rows and columns in a single scan.

Finally, mouse hover info was also added, that reports the i, j
pixel coordinates and the intensity.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>