Skip over background frames when counting files
We now completely skip over background frames when counting files.
Thus, we don't even look for files with matching indices for background
frames.

That means that if some beginning files are missing, you can also just
increase the number of background frames to skip over them.

This also modifies the prefix identifying code to automatically
select the prefix with the most files.

This was necessary because we saw there are sometimes other files
from testing that are inside the data directory that have a matching
prefix, but they are not the data itself.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>