Initial pass for projects and a project navigator
This adds a structure for projects like the following:

1. Projects
2. Sections
3. Series

A project manager is at the root and manages the projects. Each
project contains sections, and each section contains series. This
organizational structure will be used for organizing files/series.
All of these classes are serializable, and able to be saved to a
file in a json-like structure.

A project navigator was also added, whereby the user may navigate
within this structure to a particular series. Double-clicking on the
series opens the series in the PolyLaue application.

The project navigator is functioning well, but is also in an R&D state.
We may determine there is a better user interface design for some things.
But it is also highly customizable and will hopefully be easy to implement
any interface design changes.

The project manager structure is saved to the file system every time a
change is made, so it is persistent between sessions.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>