Change series on scan number change if needed
This adds support for changing series when a scan number change was
requested (if needed).

For instance, if the user requests to change to the next scan number,
but we are on the last scan number for the current series, the "Section"
(which is the parent of the series) will be checked to see if there is
an series that supports that scan number. If so, then that series will
be loaded and become the current series.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>