Do not hold file handle to external reflections
Users want to modify the HDF5 file while the application is running,
and on Windows, they cannot do this if the file is held open. Only
open the file when we want to read something.

Maybe this makes it run slower, but it's not that noticeable, and
it allows users to edit the file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>