Add ability to select and save indexing points
This adds the ability to select and save indexing points from the
"Indexing"->"Select Points" menu.

When this action is triggered, the user may left-click on the canvas
to draw points on the image, or right-click to remove points.

Once the user has finished selecting points, and clicks "OK", a file
dialog appears allowing the user to specify a directory in which the
file will be saved.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>