Prevent reflections Open button from having focus
On Windows, if a user edits a field in the reflections editor (such as
the offset), and then presses enter, the "Open External Reflections"
button gets triggered. For some reason, it appears to have focus at the
same time as the widget that is being edited. I'm wondering if this is
a bug with Qt, because only one widget should be able to have focus at
a time.

So far, the only way I could prevent it is by not allowing that button
to obtain the focus at all. This fixes the issue. It's probably okay that
the button cannot get focus, because most people would open it by clicking
on it anyways...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>