Add automatic saving for point selection
The user is now able to select from a combo box the type of point
selection that they are performing.

If "Indexing" or "Refinement" are chosen, then "indexing.xy" or
"refinement.xy" are saved automatically in the root level of the
project directory.

If "Arbitrary File" is selected, then the user chooses where the
file is saved.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>