Use one crystal at a time in Prediction Matcher
We want to allow only one crystal to be selected at a time.

This allows other crystals to still be displayed while we are matching,
so we can see if other crystals have already been predicted.

It also only overwrites any predictions for the same crystal ID within
the table.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>