Add ability to set background image for a series
This adds the basic ability to set the background image for a series.

This will be improved upon further, such as adding the ability to
set the current image in the display to be the background image for
either the series or a whole section.

We will also add a checkbox to enable/disable background subtraction.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>