Add additional background subtraction settings
This adds the ability to right-click the image view and set the current
image as the background for either the series or the entire section.

It also adds the ability to auto-level the histogram.

And, it adds the ability to toggle on/off background subtraction for the
view (it is enabled by default).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>