Make parents a requirement for most model classes
This simplifies some of the code logic where we previously had to
check whether a model class had a parent.

There aren't any situations currently, or foreseen, where we will
need to work with model classes that don't have parents. So make
parents a requirement.

This also removes the "File" menu from the main window menu bar,
because the only option was to open a series (which was created without
a parent), and this was unused.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>