Add "activate" and "clear" buttons to burn dialog
There is now a checkbox which indicates whether burning is activated.
The user must check this in order for the burn function to be applied.
It is automatically unchecked when certain things occur, such as changing
frames, so that old reflections are not accidentally overwritten.

There is also a "clear" button which clears all reflections for the
current frame from the HDF5 file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>