Add support for live acquisition
If enabled, the filesystem is constantly checked to see if enough data
exists to create a new scan within the currently viewed series.

If enough data exists, the new scan will be added automatically and
the currently viewed scan number changed to be the new scan number.

We currently check for new data every third of a second. This still
allows for interaction in the application (although it might be a tad
slower). It was not noticeably slower for me.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>