Add angular shift tracking for crystals
Now, when track() is performed, instead of creating a whole new
crystal with its own ABC matrix, we instead determine the angular
shift matrix and store that matrix in the HDF5 file, for that crystal.

The checkbox "Apply angular shift?" in the burn dialog is then used
for applying this angular shift.

Users have the options to re-use angular shifts from the first
crystal when indexing a new crystal, and apply the angular shift
to all crystals when performing tracking.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>