Allow track to overwrite ABC matrix
This adds an option to overwrite the ABC matrix within track.

If checked, the ABC matrix is overwritten with the newly discovered
ABC matrix, and all angular shift matrices are recomputed.

This also removes the old options for copying angular shift arrays
between crystals. We'll have a better way to do that soon.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>