Add a "saturation check" in the image context menu
This allows the user to select a saturation level, which is then used
to check for any pixels above that value.

If any saturated pixels are found, they are marked with triangles and
the colormap ranges changed to show only those saturated pixels.

The user's selected saturation level is remembered so that it is the
new default when the saturation check is performed again.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>