Allow points to be picked within predictions
The `event.pos()` actually returns a different type of position,
depending on whether the image was clicked or an artist/plot item
was clicked (such as a reflection prediction).

However, `event.scenePos()` always returns the scene position, which
is what we want.

After using that, picking points within predictions now works properly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>