Force replace ABC matrix if tracking original one
If a user runs `Track Orientation` on the original scan number that
was used to produce an ABC matrix (via `find()`), we now force the
user to replace that matrix with the new one.

This makes sense as the previous behavior of adding an angular
shift did not make sense, considering it would have been an angular
shift to itself.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>