Add option to use nearest ABC matrix to track
This allows a user to utilize the nearest ABC matrix as the starting
ABC matrix, rather than using the original ABC matrix. This can make
tracking significantly faster and is necessary in some situations
where the angle between the original ABC matrix and the new one would
be too high.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>