Add ability to clear new crystal from find
This allows the new crystal from Find Orientations to be removed and
deleted.

This is especially helpful if the reflections look incorrect, and you
want to start the indexing workflow over. Before, you would always have
to create a new crystal and keep every crystal. But now, you can remove
the one that was recently added.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>