Perform find/track point picking in single dialog
This moves the point picking for find/track into the find/track dialogs,
so that point picking is not a prior step and is instead combined with
the find/track step.

This is highly convenient, especially when the results appear incorrect
from find/track and you need to re-pick points. Now instead of starting
over from scratch, the user can just re-pick them in the same dialog.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>